+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## [0.2.94] - 2018-09-02
+### Added
+- The `redact()` function and `redact` option on `attachment`s.
+### Changed
+- The `template` and `table` blocks now become `DALazyTemplate`
+  objects rather than `DATemplate` objects.  Their contents are
+  evaluated at the time they are reduced to text, not at the time they
+  are created.
+### Fixed
+- Problem with `currency()` and other functions when used from a `docx
+  template file` where undefined variables did not raise exceptions.
+
 ## [0.2.93] - 2018-08-31
 ### Fixed
 - Problem with `DATemplate` variables with string indices.